requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Mail==0.9.1
email-validator==2.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
